Add 0-malloc_checked.c files

+++ 0x0C-more_malloc_free/0-malloc_checked.c
@@ -8,7 +8,8 @@
  */
 void *malloc_checked(unsigned int b)
 {
-	float *p;
+	int *p;
+
 	p = malloc(b);
 	if (p == '\0')
 		exit (98);
